dominium-56 Command Gateway und Query Gateway zur Wiederverwendung vorbereiten Refactoring:
