+++ ui/frontend/src/main/src/app/inventur/inventar/inventar.component.spec.ts
@@ -8,10 +8,10 @@ import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { of } from 'rxjs/observable/of';
 import { AppCovalentModuleModule } from '../../shared/app-covalent-module.module';
 import { AppMaterialModule } from '../../shared/app-material-module';
-import { INVENTUR_INITIAL_STATE} from '../inventur.redux';
+import { INVENTUR_INITIAL_STATE} from '../shared/inventur.redux';
 import { BilanzComponent } from '../bilanz/bilanz.component';
-import { Inventar } from '../inventar';
-import { InventurState } from '../inventur.redux';
+import { Inventar } from '../shared/inventar';
+import { InventurState } from '../shared/inventur.redux';
 import { InventurService } from '../inventur.service';
 import { InventarComponent } from './inventar.component';
 import { PositionComponent } from './position/position.component';
