+++ ui/frontend/src/main/src/app/inventur/inventar/inventar.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { InventurState } from '../inventur.redux';
-import { Inventar } from '../inventar';
+import { InventurState } from '../shared/inventur.redux';
+import { Inventar } from '../shared/inventar';
 import { Observable } from 'rxjs/Observable';
 
 import { InventurService } from '../inventur.service';
