+++ ui/frontend/src/main/src/app/inventur/shared/command-type.ts
