+++ ui/frontend/src/main/src/app/inventur/shared/inventar.ts
