+++ ui/frontend/src/main/src/app/inventur/shared/inventarposition.ts
