+++ ui/frontend/src/main/src/app/inventur/shared/query-type.ts
@@ -1,13 +1,7 @@
-import { Query } from '../shared/query-gateway/query-decorator';
-
-// const haushaltsbuch = 'com.github.haschi.dominium.haushaltsbuch.core.model.queries';
+import { Query } from '../../shared/query-gateway/query-decorator';
 
 export class QueryType {
     static readonly KeinQuery = '';
-
-    // @Query(haushaltsbuch)
     static readonly LeseInventar = 'com.github.haschi.dominium.haushaltsbuch.core.model.queries.LeseInventar';
-
-    // @Query(haushaltsbuch)
     static readonly LeseEroeffnungsbilanz = 'com.github.haschi.dominium.haushaltsbuch.core.model.queries.LeseEröffnungsbilanz';
 }
