+++ ui/frontend/src/main/src/app/inventur/shared/result-type.ts
